initial test of creating an apihandler to separate out the API business logic from the webserver
