+++ app/webserver.cpp
@@ -37,71 +37,6 @@ extern "C" {
 
 #include <fileMap.h>
 
-struct TcpPcbStats
-{
-	uint16_t active_total{0};
-	uint16_t established{0};
-	uint16_t syn_sent{0};
-	uint16_t syn_rcvd{0};
-	uint16_t fin_wait_1{0};
-	uint16_t fin_wait_2{0};
-	uint16_t close_wait{0};
-	uint16_t closing{0};
-	uint16_t last_ack{0};
-	uint16_t time_wait{0};
-	uint16_t closed{0};
-};
-
-static TcpPcbStats getTcpPcbStats()
-{
-	TcpPcbStats stats;
-#if defined(ARCH_ESP8266) || defined(ARCH_ESP32)
-	for(const tcp_pcb* pcb = tcp_active_pcbs; pcb != nullptr; pcb = pcb->next) {
-		++stats.active_total;
-		switch(pcb->state) {
-		case ESTABLISHED:
-			++stats.established;
-			break;
-		case SYN_SENT:
-			++stats.syn_sent;
-			break;
-		case SYN_RCVD:
-			++stats.syn_rcvd;
-			break;
-		case FIN_WAIT_1:
-			++stats.fin_wait_1;
-			break;
-		case FIN_WAIT_2:
-			++stats.fin_wait_2;
-			break;
-		case CLOSE_WAIT:
-			++stats.close_wait;
-			break;
-		case CLOSING:
-			++stats.closing;
-			break;
-		case LAST_ACK:
-			++stats.last_ack;
-			break;
-		case TIME_WAIT:
-			++stats.time_wait;
-			break;
-		case CLOSED:
-			++stats.closed;
-			break;
-		default:
-			break;
-		}
-	}
-#endif
-	return stats;
-}
-
-static uint16_t getActiveTcpConnectionCount()
-{
-	return getTcpPcbStats().active_total;
-}
-
 //#define NOCACHE
 
 
@@ -827,150 +762,25 @@ void ApplicationWebserver::onInfo(HttpRequest& request, HttpResponse& response){
 
 	auto stream = std::make_unique<JsonObjectStream>(2048);
 	JsonObject data = stream->getRoot();
-	debug_i("onInfo request v=%s", request.getQueryParameter(F("V")).c_str());
-	if(request.getQueryParameter(F("V")) == "2"||request.getQueryParameter(F("v")) == "2"  ){
-		debug_i("onInfo v2");
-		addInfoFields(data);
-		const auto tcpStats = getTcpPcbStats();
 
-		#ifndef SMING_RELEASE
-	
-		const char* preNetState = "unknown";
-		switch(app.syslogPreNetState()) {
-			case UdpSyslogStream::PreNetState::Buffering:
-				preNetState = "buffering";
-				break;
-			case UdpSyslogStream::PreNetState::Draining:
-				preNetState = "draining";
-				break;
-			case UdpSyslogStream::PreNetState::Done:
-				preNetState = "done";
-				break;
-		}
-		JsonObject debug = data.createNestedObject(F("debug"));
-		debug[F("syslog_pre_net_state")] = preNetState;
-		debug[F("http_active_connections")] = activeClients;
-		debug[F("websocket_connections")] = webSockets.size();
-		debug[F("eventserver_clients")] = app.eventserver.activeClients;
-		debug[F("syslog_pre_net_buffer_allocated")] = app.udpSyslogStream.preNetBufferAllocated();
-		debug[F("syslog_pre_net_encoder_allocated")] = app.udpSyslogStream.preNetEncoderAllocated();
-		debug[F("syslog_pre_net_buffer_capacity")] = app.udpSyslogStream.preNetBufferCapacity();
-		debug[F("syslog_pre_net_buffer_used")] = app.udpSyslogStream.preNetBufferUsed();
-		debug[F("syslog_pre_net_buffer_frames")] = app.udpSyslogStream.preNetBufferedFrames();
-		debug[F("syslog_pre_net_buffer_evicted")] = app.udpSyslogStream.preNetEvictedFrames();
-		debug[F("tcp_pcb_size")] = sizeof(tcp_pcb);
-		debug[F("tcp_active_estimated_bytes")] = static_cast<uint32_t>(tcpStats.active_total) * sizeof(tcp_pcb);
+	if(!app.api) {
+		setCorsHeaders(response);
+		response.code = HTTP_STATUS_SERVICE_UNAVAILABLE;
+		response.setContentType(MIME_TEXT);
+		response.sendString(F("api not initialized"));
+		return;
+	}
 
-		#endif
-		
-		debug_i("onInfo adding rgbww object");
-		
-		JsonObject rgbww = data.createNestedObject(F("rgbww"));
-		rgbww[F("version")] = RGBWW_VERSION;
-		rgbww[F("queuesize")] = RGBWW_ANIMATIONQSIZE;
+	StaticJsonDocument<64> paramsDoc;
+	JsonObject params = paramsDoc.to<JsonObject>();
+	params[F("V")] = request.getQueryParameter(F("V"));
+	params[F("v")] = request.getQueryParameter(F("v"));
 
-		debug_i("onInfo adding connection object");
-		JsonObject con = data.createNestedObject(F("connection"));
-		con[F("connected")] = WifiStation.isConnected();
-		if(WifiStation.isConnected()) {
-			con[F("ssid")] = WifiStation.getSSID();
-			con[F("dhcp")] = WifiStation.isEnabledDHCP();
-			con[F("ip")] = WifiStation.getIP().toString();
-			con[F("netmask")] = WifiStation.getNetworkMask().toString();
-			con[F("gateway")] = WifiStation.getNetworkGateway().toString();
-			con[F("mac")] = WifiStation.getMAC();
-			con[F("rssi")] = WifiStation.getRssi();
+	if(!app.api->dispatch(F("info"), params, data)) {
+		sendApiCode(response, API_CODES::API_BAD_REQUEST, F("unsupported api method"));
+		return;
+	}
 
-			JsonObject net = data.createNestedObject(F("network"));
-			net[F("tcp_connections")] = getActiveTcpConnectionCount();
-			net[F("tcp_active")]= tcpStats.active_total;
-			net[F("tcp_established")] = tcpStats.established;
-			net[F("tcp_syn_sent")] = tcpStats.syn_sent;
-			net[F("tcp_syn_rcvd")] = tcpStats.syn_rcvd;
-			net[F("tcp_fin_wait_1")] = tcpStats.fin_wait_1;
-			net[F("tcp_fin_wait_2")] = tcpStats.fin_wait_2;
-			net[F("tcp_close_wait")] = tcpStats.close_wait;
-			net[F("tcp_closing")] = tcpStats.closing;
-			net[F("tcp_last_ack")] = tcpStats.last_ack;
-			net[F("tcp_time_wait")] = tcpStats.time_wait;
-			net[F("tcp_closed")] = tcpStats.closed;
-		}
-		// Skip ConfigDB reads during OTA — they consume heap and flash I/O we can't afford
-		debug_i("onInfo adding OTA object");
-		if(!app.ota.isProccessing()) {
-			AppConfig::Network network(*app.cfg);
-			JsonObject mqtt=data.createNestedObject(F("mqtt"));
-			if(network.mqtt.getEnabled() && !app.mqttclient.isRunning()) {
-				mqtt[F("status")] = F("configured but not running");
-			} else if(network.mqtt.getEnabled() && app.mqttclient.isRunning()) {
-				mqtt[F("status")] = F("running");
-			} else {
-				mqtt[F("status")] = F("disabled");
-			}
-			mqtt[F("enabled")] = network.mqtt.getEnabled();
-			mqtt[F("broker")] = network.mqtt.getServer();
-			mqtt[F("topic")] = network.mqtt.getTopicBase();
-			if(network.mqtt.homeassistant.getEnable())
-			{
-				JsonObject ha = data.createNestedObject("homeassistant");
-				ha[F("enabled")] = network.mqtt.homeassistant.getEnable();
-				ha[F("discovery_prefix")] = network.mqtt.homeassistant.getDiscoveryPrefix();
-				ha[F("Node ID")]= network.mqtt.homeassistant.getNodeId();
-			}
-		}
-
-		if(app.ota.isProccessing()) {
-			JsonObject ota=data.createNestedObject(F("ota"));
-			ota[F("status")] = F("in progress");
-		}
-
-	}else{
-		debug_i("old style onInfo");	
-		#ifdef ARCH_ESP8266
-			if(app.ota.isProccessing()) {
-				sendApiCode(response, API_CODES::API_UPDATE_IN_PROGRESS);
-				return;
-			}
-		#endif
-
-			data[F("deviceid")] = system_get_chip_id();
-		#if defined(ARCH_ESP8266) || defined(ARCH_ESP32)
-			data[F("current_rom")] = String(app.ota.getRomPartition().name());
-		#endif
-			data[F("git_version")] = fw_git_version;
-			data[F("build_type")] = BUILD_TYPE;
-			data[F("git_date")] = fw_git_date;
-			data[F("webapp_version")] = WEBAPP_VERSION;
-			data[F("sming")] = SMING_VERSION;
-			data[F("event_num_clients")] = app.eventserver.activeClients;
-			data[F("uptime")] = app.getUptime();
-			data[F("heap_free")] = app.getFreeHeapSize();
-			data[F("soc")]=SOC;
-			
-			/*
-			FileSystem::Info fsInfo;
-			app.getinfo(fsInfo);
-			JsonObject FS=data.createNestedObject("fs");
-			FS[F("mounted")] = fsInfo.partition?"true":"false";
-			FS[F("size")] = fsInfo.total;
-			FS[F("used")] = fsInfo.used;
-			FS[F("available")] = fsInfo.freeSpace;
-		*/
-			JsonObject rgbww = data.createNestedObject("rgbww");
-			rgbww[F("version")] = RGBWW_VERSION;
-			rgbww[F("queuesize")] = RGBWW_ANIMATIONQSIZE;
-
-			JsonObject con = data.createNestedObject("connection");
-			con[F("connected")] = WifiStation.isConnected();
-			con[F("ssid")] = WifiStation.getSSID();
-			con[F("dhcp")] = WifiStation.isEnabledDHCP();
-			con[F("ip")] = WifiStation.getIP().toString();
-			con[F("netmask")] = WifiStation.getNetworkMask().toString();
-			con[F("gateway")] = WifiStation.getNetworkGateway().toString();
-			con[F("mac")] = WifiStation.getMAC();
-			//con[F("mdnshostname")] = app.cfg.network.connection.mdnshostname.c_str();
-
-		}
 		sendApiResponse(response, stream.release());
 
 }
