+++ include/apihandler.h
@@ -1,0 +1,31 @@
+/**
+ * @author  Peter Jakobs http://github.com/pljakobs
+ * 
+ * @section LICENSE
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation; either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details at
+ * https://www.gnu.org/copyleft/gpl.html
+ */
+
+#pragma once
+
+#include <ArduinoJson.h>
+
+class Api {
+public:
+	Api() = default;
+	~Api() = default;
+
+	bool dispatch(const String& method, const JsonObject& params, JsonObject& out);
+
+private:
+	bool handleInfo(const JsonObject& params, JsonObject& out);
+};
